move index.js to esm module and support cjs with fallback
